hotel-california/requirements.txt
alembic==1.16.2
altair==4.2.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
appdirs==1.4.4
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arro3-core==0.5.1
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
avro==1.11.3
babel==2.17.0
backoff==2.2.1
beautifulsoup4==4.13.4
binaryornot==0.4.4
bleach==6.2.0
blinker==1.9.0
blosc2==3.4.0
boto3==1.38.45
botocore==1.38.45
bs4==0.0.1
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
colorlog==6.9.0
comm==0.2.2
contourpy==1.3.2
cookiecutter==2.6.0
cryptography==45.0.4
cycler==0.12.1
dacite==1.9.2
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
databricks-cli==0.18.0
databricks-sdk==0.57.0
db-dtypes==1.4.3
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
deltalake==1.0.2
Deprecated==1.2.18
distro==1.9.0
docker==7.1.0
dynaconf==3.2.11
entrypoints==0.4
et_xmlfile==2.0.0
executing==2.2.0
fastapi==0.115.14
fastjsonschema==2.21.1
Flask==2.3.3
fonttools==4.58.4
fqdn==1.5.1
fsspec==2025.5.1
furl==2.1.4
geopandas==1.0.1
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.25.1
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-cloud-bigquery==3.34.0
google-cloud-core==2.4.3
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
great-expectations==0.15.50
greenlet==3.2.3
grpcio==1.73.0
grpcio-status==1.56.2
h11==0.16.0
holidays==0.68
hopsworks==4.2.6
hopsworks-aiomysql==0.2.1
htmlmin==0.1.12
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ImageHash==4.3.1
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
itsdangerous==2.2.0
javaobj-py3==0.4.4
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.5.1
json5==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kedro==0.19.14
kedro-datasets==7.0.0
kedro-telemetry==0.6.3
kedro_mlflow==0.14.4
kiwisolver==1.4.8
kmodes==0.12.2
langdetect==1.0.9
lazy_loader==0.4
lightgbm==4.6.0
llvmlite==0.44.0
lxml==4.9.4
makefun==1.16.0
Mako==1.3.10
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
missingno==0.5.2
mistune==3.1.3
mlflow==2.22.1
mlflow-skinny==2.22.1
mock==5.2.0
monotonic==1.6
more-itertools==10.7.0
msgpack==1.1.1
multimethod==1.9
narwhals==1.38.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
ndindex==1.10.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.8.1
notebook==7.4.3
notebook_shim==0.2.4
numba==0.61.2
numexpr==2.11.0
numpy==1.26.4
oauthlib==3.3.1
omegaconf==2.3.0
openai==1.79.0
openpyxl==3.1.5
opensearch-py==2.4.2
opentelemetry-api==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
optuna==4.4.0
orderedmultidict==1.0.1
overrides==7.7.0
packaging==24.2
pandas==2.1.4
pandas-gbq==0.29.1
pandocfilters==1.5.1
parse==1.20.2
parso==0.8.4
patsy==1.0.1
phik==0.12.4
pillow==11.2.1
platformdirs==4.3.8
plotly==6.0.0
pluggy==1.6.0
posthog==3.25.0
pre_commit_hooks==5.0.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==4.25.4
psutil==7.0.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodomex==3.23.0
pydantic==1.10.12
pydantic-settings==2.0.3
pydantic_core==2.33.2
pydata-google-auth==1.9.1
pydeck==0.9.1
Pygments==2.19.1
pyhumps==1.6.1
pyjks==20.0.0
PyJWT==2.10.1
PyMySQL==1.1.1
pynndescent==0.5.13
pyodbc==5.2.0
pyogrio==0.10.0
pyparsing==3.2.3
pyproj==3.7.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-json-logger==3.3.0
python-slugify==8.0.4
pytoolconfig==1.3.1
pytz==2025.2
PyWavelets==1.8.0
pywin32==310
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==27.0.0
querystring-parser==1.2.4
referencing==0.36.2
regex==2023.10.3
requests==2.32.4
requests-oauthlib==2.0.0
retrying==1.4.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rope==1.13.0
rpds-py==0.25.1
rsa==4.9.1
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.12
s3transfer==0.13.0
scikit-learn==1.7.0
scipy==1.16.0
seaborn==0.13.2
Send2Trash==1.8.3
shap==0.48.0
shapely==2.1.1
six==1.17.0
slicer==0.0.8
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.29
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.46.2
statsmodels==0.14.4
streamlit==1.45.0
streamlit-option-menu==0.3.6
tables==3.10.2
tabulate==0.9.0
tenacity==9.1.2
terminado==0.18.1
text-unidecode==1.3
threadpoolctl==3.6.0
tinycss2==1.4.0
toml==0.10.2
toolz==1.0.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
twofish==0.3.0
typeguard==4.4.2
types-python-dateutil==2.9.0.20250516
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
umap==0.1.1
umap-learn==0.5.7
uri-template==1.3.0
urllib3==1.26.20
uv==0.7.11
uvicorn==0.34.3
vaderSentiment==3.3.2
validators==0.22.0
visions==0.7.6
waitress==3.0.2
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.6
widgetsnbextension==4.0.14
wordcloud==1.9.4
wrapt==1.17.2
xgboost==1.3.3
ydata-profiling==4.12.2
yellowbrick==1.5
zipp==3.23.0
